requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
psycopg2-binary
python-dotenv
bcrypt~=4.1.3
datetime~=5.5
python-dotenv==1.0.1
pre-commit==3.7.1
